feat: edit plugins organization
